Updated path of diffrot
